requirements.txt
python-telegram-bot==20.0
pymongo==4.0
